pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Teacher Dashboard API for managing slides and content"
authors = [
    {name = "Ahsan", email = "ahs462agk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn[standard]>=0.15.0",
    "python-dotenv>=0.19.0",
    "python-multipart>=0.0.5",
    "langchain-google-genai>=0.1.0",
    "langchain-community>=0.0.10",
    "pinecone-client>=3.0.0",
    "python-magic-bin; sys_platform == 'win32'",
    "python-magic; sys_platform != 'win32'",
    "pypdf>=3.0.0",
    "pinecone (>=7.3.0,<8.0.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
